VISA-ASK-SYSTEM/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
requests==2.31.0
openai==1.28.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.4.2
pandas==2.1.1
openpyxl==3.1.2
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.0 